+++ core/identity-hub-did/build.gradle.kts
@@ -8,6 +8,8 @@ dependencies {
 
     implementation(libs.edc.core.connector) // for the reflection-based query resolver
     implementation(libs.edc.common.crypto)
+    implementation(libs.edc.lib.store)
+    implementation(libs.edc.lib.query)
 
     testImplementation(libs.edc.junit)
     testImplementation(libs.edc.ext.jsonld)
