+++ core/identity-hub-did/src/test/java/org/eclipse/edc/identityhub/did/defaults/InMemoryDidResourceStoreTest.java
@@ -14,9 +14,9 @@
 
 package org.eclipse.edc.identityhub.did.defaults;
 
-import org.eclipse.edc.connector.core.store.CriterionOperatorRegistryImpl;
 import org.eclipse.edc.identithub.did.spi.store.DidResourceStore;
 import org.eclipse.edc.identityhub.did.store.test.DidResourceStoreTestBase;
+import org.eclipse.edc.query.CriterionOperatorRegistryImpl;
 
 class InMemoryDidResourceStoreTest extends DidResourceStoreTestBase {
